+++ tests/deps.ts
@@ -1,0 +1,4 @@
+export {
+  assert,
+  assertStrictEquals,
+} from "https://deno.land/std@0.69.0/testing/asserts.ts";