+++ types/vendor.ts
@@ -1,0 +1,3 @@
+import type { vendors } from "../vendors.ts";
+
+export type Vendor = typeof vendors[number];