+++ vendors.ts
@@ -1,0 +1,19 @@
+// My OCD wishes it could be a type of `Vendor[]` but ts just doesn't work that way :(
+export const vendors = [
+  "ah",
+  "apple",
+  "atsc",
+  "epub",
+  "fx",
+  "hp",
+  "khtml",
+  "moz",
+  "ms",
+  "o",
+  "rim",
+  "ro",
+  "tc",
+  "wap",
+  "webkit",
+  "xv",
+] as const;