+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## 0.2.0 - [2020-09-17]
+
+- new `Vendor` type
+- new `isVendor()` for checking if something is a vendor
+- added `./bundle/web/deno-vendors.js` bundle for the web
+- updated `mod.ts`
+- updated tests
+
 ## 0.1.0 - [2020-08-07]
 
 - first release